// src/orders/interfaces/insert-cart-response.interface.ts
export default interface InsertCartResponse {
  id: string;
  protocol: string;
  service_id: number;
  agency_id?: number;
  contract: string;
  service_code?: number;
  quote: number;
  price: number;
  coupon: null;
  discount: number;
  delivery_min: number;
  delivery_max: number;
  status: string;
  reminder: null;
  insurance_value: number;
  weight: null;
  width: null;
  height: null;
  length: null;
  diameter: null;
  format: string;
  billed_weight: number;
  receipt: boolean;
  own_hand: boolean;
  collect: boolean;
  collect_scheduled_at: null;
  reverse: boolean;
  non_commercial: boolean;
  authorization_code: null;
  tracking: null;
  self_tracking: null;
  delivery_receipt: null;
  additional_info: null;
  cte_key: null;
  paid_at: null;
  generated_at: null;
  posted_at: null;
  delivered_at: null;
  canceled_at: null;
  suspended_at: null;
  expired_at: null;
  created_at: string;
  updated_at: string;
  parse_pi_at: null;
  products: [[Object]];
  volumes: [[Object]];
  tags: [[Object]];
}